moved map positioning control to App
