+++ src/App.js
@@ -1,8 +1,53 @@
-import React from 'react'
+import React, { useState, useEffect } from 'react'
 import MapView from './components/MapView'
 import { MapProvider } from 'react-map-gl'
+import defaultLocations from './data/defaultLocations'
+
+function getCoords () {
+	return new Promise( ( resolve, reject ) => {
+		navigator.geolocation.getCurrentPosition(
+			pos => resolve( {
+				longitude: pos.coords.longitude,
+				latitude: pos.coords.latitude,
+				zoom: 9.5
+			} ),
+			err => reject( err )
+		)
+	} )
+}
+
+// Random interger between 0 and 9.
+// Used to pick one of ten default locations.
+const int = Math.floor( Math.random() * 10 )
 
 export default function App () {
+	// Initialize viewState (map center position and zoom value) as null.
+	// If viewState is null, the map will not render.
+	const [viewState, setViewState] = useState( null )
+
+	// Get position fires just once, on load.
+	useEffect( () => {
+		getPosition()
+	}, [] )
+
+	// Gets initial position from the browser. 
+	// This function is called in a useEffect.
+	// Uses a default position if the browser does not have permission.
+	async function getPosition () {
+		try {
+			const coords = await getCoords()
+			setViewState( coords )
+		}
+		catch ( err ) {
+			console.log( err.message )
+			setViewState( {
+				longitude: defaultLocations[int][0],
+				latitude: defaultLocations[int][1],
+				zoom: 9.5
+			} )
+		}
+	}
+
 	return (
 		<>
 			<header>
@@ -31,7 +76,7 @@ export default function App () {
 			</header>
 			<main id="main">
 				<MapProvider>
-					<MapView />
+					<MapView viewState={ viewState } setViewState={ setViewState } />
 				</MapProvider>
 			</main>
 			<footer className="footer mt-auto py-3 px-2 bg-light">
